1) Added extensions for UIButton so that all buttons are now rounded and floating 2) Increase all the buttons size 2) Set the leaderboard label to be shorter 3) Added the shadow to pencils when pressed

+++ Engagingu/Controller/AnagramViewController.swift
@@ -22,7 +22,7 @@ class AnagramViewController: UIViewController, UITextFieldDelegate {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-        
+        submitBtn.float()
         // Create border for input box
         let myColor = UIColor.black
         wordInput.layer.borderWidth = 1.0
